pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-nexus-agent"
version = "0.3.0"
description = "Nexus — MCP server for Home Assistant; 202 tools across 21 domains for AI control of your smart home"
authors = [{ name = "Łukasz Żochowski" }]
license = { text = "MIT" }
readme = "README.md"
keywords = ["home-assistant", "mcp", "model-context-protocol", "smart-home", "ai", "claude", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.40",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
nexus = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
